+++ owl_sql/test/pg_sample_test.dart
@@ -8,11 +8,11 @@ import 'golden/pg_sample.g.dart';
 Future main() async {
   group('pg_scan', () {
     PostgreSQLConnection conn;
-    final table = new SampleTable('stbl', schema: 'test_sample');
+    final table = SampleTable('stbl', schema: 'test_sample');
 
     setUpAll(() async {
       // docker run --rm -it -p 5432:5432 postgres:11.1
-      conn = new PostgreSQLConnection('localhost', 5432, 'postgres',
+      conn = PostgreSQLConnection('localhost', 5432, 'postgres',
           username: 'postgres', password: 'postgres');
       await conn.open();
       await conn.execute('CREATE SCHEMA IF NOT EXISTS test_sample;');
@@ -26,7 +26,7 @@ Future main() async {
     test('populate data', () async {
       await table.insert(
           conn,
-          new SampleRow(
+          SampleRow(
             textCol: 'id-value',
             bigintCol: 1 << 62,
             smallintCol: -135,
@@ -37,7 +37,7 @@ Future main() async {
               'a': 'a1',
               'b': [1, 2, 3]
             },
-            timestampCol: new DateTime(2000, 01, 02, 03, 04, 05).toUtc(),
+            timestampCol: DateTime(2000, 01, 02, 03, 04, 05).toUtc(),
             uuidCol: '00112233-4455-6677-8899-aabbccddeeff',
           ));
 
@@ -57,7 +57,7 @@ Future main() async {
         'a': 'a1',
         'b': [1, 2, 3]
       });
-      expect(row.timestampCol, new DateTime(2000, 01, 02, 03, 04, 05).toUtc());
+      expect(row.timestampCol, DateTime(2000, 01, 02, 03, 04, 05).toUtc());
       expect(row.uuidCol, '00112233-4455-6677-8899-aabbccddeeff');
     });
 
@@ -67,17 +67,17 @@ Future main() async {
         expect(rows, hasLength(1));
       }
 
-      await run(new SampleFilter()..bigintCol$equalsTo(4611686018427387904));
-      await run(new SampleFilter()..bigintCol$greaterThan(4611686018427387903));
-      await run(new SampleFilter()..smallintCol$equalsTo(-135));
-      await run(new SampleFilter()..smallintCol$lessThan(-134));
-      await run(new SampleFilter()..booleanCol$equalsTo(true));
-      await run(new SampleFilter()..byteaCol$greaterThan([1, 1]));
-      await run(new SampleFilter()..doubleCol$greaterThan(4.4));
-      await run(new SampleFilter()..jsonbCol$matches({'a': 'a1'}));
-      await run(new SampleFilter()
-        ..timestampCol$greaterThan(new DateTime(1999, 01, 01)));
-      await run(new SampleFilter()
+      await run(SampleFilter()..bigintCol$equalsTo(4611686018427387904));
+      await run(SampleFilter()..bigintCol$greaterThan(4611686018427387903));
+      await run(SampleFilter()..smallintCol$equalsTo(-135));
+      await run(SampleFilter()..smallintCol$lessThan(-134));
+      await run(SampleFilter()..booleanCol$equalsTo(true));
+      await run(SampleFilter()..byteaCol$greaterThan([1, 1]));
+      await run(SampleFilter()..doubleCol$greaterThan(4.4));
+      await run(SampleFilter()..jsonbCol$matches({'a': 'a1'}));
+      await run(
+          SampleFilter()..timestampCol$greaterThan(DateTime(1999, 01, 01)));
+      await run(SampleFilter()
         ..uuidCol$greaterThan('00112233-0000-6677-8899-aabbccddeeff'));
     });
 
@@ -85,11 +85,11 @@ Future main() async {
       await table.update(
           conn,
           'id-value',
-          new SampleUpdate()
+          SampleUpdate()
             ..smallintCol(135)
             ..byteaCol([2, 3])
             ..jsonbCol({'x': 1})
-            ..timestampCol(new DateTime(2002, 02, 03, 04, 05, 06).toUtc())
+            ..timestampCol(DateTime(2002, 02, 03, 04, 05, 06).toUtc())
             ..uuidCol('22110033-4455-6677-8899-aabbccddeeff'));
     });
 
@@ -102,7 +102,7 @@ Future main() async {
       expect(row.byteaCol, [2, 3]);
       expect(row.doubleCol, 4.5);
       expect(row.jsonbCol, {'x': 1});
-      expect(row.timestampCol, new DateTime(2002, 02, 03, 04, 05, 06).toUtc());
+      expect(row.timestampCol, DateTime(2002, 02, 03, 04, 05, 06).toUtc());
       expect(row.uuidCol, '22110033-4455-6677-8899-aabbccddeeff');
     });
   });